[WIP] fix to adapt activity lifecycle
